update files after comments from lesson 3
