+++ frontend/src/components/FilmCardButton/FilmCardButton.scss
@@ -1,14 +1,5 @@
 @import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700&display=swap');
 
-.buttonGroup {
-    width: 100%;
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    margin-bottom: 10px;
-    padding: 0 15px;
-  }
-
 .button {
     width: 48%;
     font: 600 12px 'Roboto', sans-serif;
