+++ frontend/src/pages/FilmCard/FilmCard.scss
@@ -1,0 +1,116 @@
+@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700&display=swap');
+@import '../../styles/variables.scss';
+
+@mixin fontFrame($weight, $size, $color) {
+    font: $weight $size 'Roboto', sans-serif;
+    color: $color;
+}
+
+.filmCard {
+    &--wrapper {
+      min-width: 151px;
+      height: 100vh;
+      display: flex;
+      flex-direction: column;
+      background-color: $templateBackgroundColor;
+    }
+    &__poster {
+      width: 100%;
+      height: 35%;
+      object-fit: cover;
+      background-color: lightgray;
+      display: flex;
+      &__image {
+        width: 100%;
+      }
+    }
+    &__infoBlock {
+      width: 100%;
+      min-height: 49.5%;
+      position: relative;
+      display: flex;
+      flex-direction: column;
+      flex: 1 0 auto;
+      border-radius: 35px 35px 0 0;
+      background-color: $templateBackgroundColor;
+      font: $generalFont;
+      color: $textColor;
+      margin-top: -50px;
+      padding: 19px 15px 0;
+      &--titleWrapper {
+        display: flex;
+        justify-content: space-between;
+        padding: 0 10px;
+      }
+      &__title {
+        width: 80%;
+        @include fontFrame(500, 25px, black);
+        text-transform: uppercase;
+        hyphens: auto;
+        margin: 0;
+      }
+      &__rate {
+        display: flex;
+        justify-content: flex-end;
+        align-self: flex-start;
+        @include fontFrame(400, 20px, rgb(223, 160, 26));
+        margin: 0 0 0 5px;
+      }
+      &--secondTitleWrapper {
+        display: flex;
+        justify-content: space-between;
+        margin: 2px 0 12px;
+        padding: 0 10px;
+      }
+      &__secondTitle {
+        width: 50%;
+        text-transform: capitalize;
+        hyphens: auto;
+        margin: 0;
+      }
+      &__year {
+        width: 50%;
+        text-align: right;
+        margin: 0;
+      }
+      &__genre {
+        text-align: center;
+        margin: 0 0 7px;
+      }
+      &--underline {
+        width: 75%;
+        color: rgb(112, 112, 112);
+        margin: 0 auto 13px;
+      }
+      &__description {
+        text-align: justify;
+        hyphens: auto;
+        margin: 0 0 10px;
+      }
+    }
+    &__footer {
+        width: 100%;
+        height: 138px;
+        flex: 0 0 auto;
+        &__buttonGroup {
+            width: 100%;
+            display: flex;
+            flex-wrap: wrap;
+            justify-content: center;
+            margin-bottom: 10px;
+            padding: 0 15px;
+        }
+        &__emptyBlock {
+            width: 100%;
+            height: 46px;
+        }
+      }
+  }
+  
+  .navigation {
+    position: fixed;
+    width: 100%;
+    height: $navigationHeight;
+    bottom: 0;
+    background-color:gray;
+  }
